MFC r342815:
Remove ipsd (IP Scan Detetor). It is unused and to my knowledge has
never been used on any platform that ipfilter has been on. However
it looks like it could be a useful utility, therefore there are plans
to make it a port one day. It lacks a man page as well.


git-svn-id: https://svn.freebsd.org/base/stable/11@343759 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f