MFC r343535: Speed up non-status operations applied to a single interface
When performing a non-status operation on a single interface, it is
not necessary for ifconfig to build a list of all addresses in the
system, sort them, then iterate through them looking for the entry for
the single interface of interest. Doing so becomes increasingly
expensive as the number of interfaces in the system grows (e.g., in a
system with 1000+ vlan(4) interfaces).

Reviewed by:	ae, kp
Sponsored by:	RG Nets
Differential Revision:	https://reviews.freebsd.org/D18919


git-svn-id: https://svn.freebsd.org/base/stable/12@344030 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f