Fix PAE modules build on i386.
Reimplement PAE version of pte_load() by copying/pasting the
atomic_load_acq_64_i586() into it definition. pmap_kextract() is defined
as inline and uses pte_load() in its body, so the pte_load() should be
available when pmap.h is included. On stable/11, the atomic inlines are
not exposed to modules.

This is a direct commit to stable/11.

Reported by:	dim
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/11@344053 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f