MFC r333365, r333471:
r333365:
We don't use f_devdata, so don't set it. Should that need to change
later, we can. This leaves ZFS as the only irregular f_devdata
user in the tree.

r333471:
Revert r333365

Even though we don't use it, it appears something else requires it to
be != 0 to work. This breaks tftp boot in loader.efi, so revert until
that can be sorted out.


git-svn-id: https://svn.freebsd.org/base/stable/11@344287 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f