MFC r343125 (by cem): ioat(4): Set __result_use_check on ioat_acquire_reserve
Even M_WAITOK callers must check for failure.  For example, if the device is
quiescing, either due to automatic error-recovery induced reset, or due to
administrative detach, the routine will return ENXIO and the acquire
reference will not be held.  So, there is no mode in which it is safe to
assume the routine succeeds without checking.


git-svn-id: https://svn.freebsd.org/base/stable/12@344400 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f