Direct commit to catch pc98 up to recent changes
I have no means of actually testing the new features on pc98, so simply mark
them as BROKEN_OPTIONS on pc98 and interested parties can test/report back
as to the functionality. This is done out of an abundance of caution.

pc98 also needed catching up on the inflate.c -> subr_inflate.c rename.


git-svn-id: https://svn.freebsd.org/base/stable/11@344421 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f