Update the stable/11 arm SoC IMAGE_SIZE to 3072M as build failures were encountered due to the target filesystem running out of space during the installkernel target.
The 3072M size is used as it is consistent with head and
stable/12 image sizes, and build failures due to a full
filesystem have been common enough to justify making the
default size the same across all supported branches in
order to avoid such surprises.

This is a direct commit to stable/11, as there are several
changes between several size bumps that are unrelated or
unnecessary in this branch.

Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/11@344442 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f