Add support to fdt_slicer for the new style partition data documented in devicetree/bindings/mtd/partition.txt.
In the old style, all the children of the device node which did not have a
compatible property were the partitions.  In the new style, there is a child
node of the device which has a compatible string of "fixed-partitions", and
its children are the individual partitions.

Also, support the read-only property by setting the corresponding slice flag.


git-svn-id: https://svn.freebsd.org/base/head@344616 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f