MFC r344473:
CID 1332000: Logically dead code in sys/dev/pms/RefTisa/tisa/sassata/sas/ini/itdio.c

A pointer is first tested for NULL. If non-NULL, another pointer is
set equal to the first. The second pointer is then checked for NULL
and an error path taken if so. This second test and the associated
path is dead code as the pointer value, having just been checked for
NULL, cannot be NULL at this point. Remove the dead code.

Sponsored by:	Dell EMC Isilon


git-svn-id: https://svn.freebsd.org/base/stable/11@344645 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f