MFC r343978:
pfctl: Fix ifa_grouplookup()

Setting the length of the request got lost in r343287, which means SIOCGIFGMEMB
gives us the required length, but does not copy the names of the group members.
As a result we don't get a correct list of group members, and 'set skip on
<ifgroup>' broke.

This produced all sorts of very unexpected results, because we would end up
applying 'set skip' to unexpected interfaces.


git-svn-id: https://svn.freebsd.org/base/stable/12@344712 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f