Allow the sector size of the disk device to be configured using hints or FDT data.  The sector size must be a multiple of the device's page size. If not configured, use the historical default of the device page size.
Setting the disk sector size to 512 or 4096 allows a variety of standard
filesystems to be used on the device.  Of course you wouldn't want to be
writing frequently to a SPI flash chip like it was a disk drive, but for
data that gets written once (or rarely) and read often, using a standard
filesystem is a nice convenient thing.


git-svn-id: https://svn.freebsd.org/base/head@344734 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f