MFC 344671: Don't assume all children of a nexus are ports.
Specifically, ccr(4) devices are also children of cxgbe nexus devices.
Rather than making assumptions about the child device's softc, walk
the list of ports from the nexus' softc to determine if a child is a
port in t4_child_location_str().  This fixes a panic when detaching a
ccr device.


git-svn-id: https://svn.freebsd.org/base/stable/12@344933 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f